fix(ui): address review findings on the connections rework
- setAgentConnections updates the grants cache optimistically so quick
  consecutive toggles compound instead of the second full-set write
  dropping the first grant (rolled back on error)
- a popup closed while the create call is in flight discards the
  half-created connection instead of leaving an Authorizing husk
- the wizard's popup-blocked full-page OAuth return stages the
  authorized connection into the draft via the callback's connection id
- derive catalogue tab counts from the tab order

refactor(ui): split use-sandbox-settings-form into focused hooks
(schema, provider staging, inherited envs, egress preview, save
orchestration) composed by a thin parent
